requirements.txt
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-Login==0.2.11
sqlalchemy-migrate==0.10.0
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
Markdown==2.6.8

ForgeryPy==0.1